todo-app-phase2/backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.14
pydantic==2.5.0
pydantic-settings==2.1.0
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.13.1
pytest==7.4.3
httpx==0.25.2